HEART DISEASE PREDICTION
Coronary artery disease is a common heart condition that affects the major blood vessels that supply the heart muscle. A buildup of fats, cholesterol and other substances in and on the artery walls usually causes coronary artery disease. This buildup is called plaque. The buildup of plaque in the arteries is called atherosclerosis (ath-ur-o-skluh-ROE-sis). Atherosclerosis reduces blood flow to the heart and other parts of the body. It can lead to a heart attack, chest pain or a stroke.